updates and fixes
- added flag for discovery trace logger
- added libp2p ping
- added libp2p NAT service + rate limit
- removed operator/peer id from node record

+++ network/records/subnets_enr.go
@@ -11,7 +11,7 @@ import (
 
 // UpdateSubnets updates subnets entry according to the given changes.
 // count is the amount of subnets, in case that the entry doesn't exist as we want to initialize it
-func UpdateSubnets(node *enode.LocalNode, count int, added []int, removed []int) error {
+func UpdateSubnets(node *enode.LocalNode, count int, added []int64, removed []int64) error {
 	subnets, err := GetSubnetsEntry(node.Node().Record())
 	if err != nil {
 		return errors.Wrap(err, "could not read subnets entry from enr")
